A crosswalk of IEP survey names

Relates IEP survey names to recommended names created by the DUWG

Usage

crosswalk

Format

A data frame with 29 columns

variableType

Category of variable of interest

variable

Name used to describe the data collected

recommendedName

Value recommended by IEP to use to describe the relevant data

recommendedUnit

Value recommended by IEP to use to describe the unit of the relevant data

notes

Freeform comments by IEP

ybfmp

Yolo Bypass Fish Monitoring Program

YBFMP Units

Units of

djfmp

Delta Juvenile Fish Monitoring Program

DJFMP Units

Units of

frp

Fish Restoration Program

FRP Units

Units of

Delta Smelt formalin preservation data

Not familiar with this

Delta boat electrofishing survey

Not familiar with this

Electrofishing Units

Units of

EDSM

Enhanced Delta Smelt Monitoring Program

EDSM Units

Units of

sls

Smelt Larval Survey

SLS units

Units of

20mm

20 mm Survey

20mm units

Units of

stn

Summer Townet Survey

STN units

Units of

fmwt

Fall Midwater Trawl

FMWT units

Units of

skt

Spring Kodiak Trawl

SKT units

Units of

sms

Suisun Marsh Study

SMS units

Units of

bs

Bay Study